week_4/requirements.txt
langchain>=0.3.0
langgraph>=0.4.8
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-groq>=0.3.0
langchain-openai>=0.2.0
streamlit>=1.39.0
python-dotenv>=1.0.0
pydantic>=2.0.0
